backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-socketio==5.11.2
requests==2.32.3
python-dotenv==1.0.1
mistralai>=1.8.1
